refactor(hub): extract ElevatedHandle from vault.ts (Phase 5 A6)
